+++ arcane/src/arcane/std/BasicReaderWriter.cc
@@ -11,6 +11,8 @@
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
+#include "arcane/std/internal/BasicReaderWriter.h"
+
 #include "arcane/utils/TraceAccessor.h"
 #include "arcane/utils/ScopedPtr.h"
 #include "arcane/utils/StringBuilder.h"
@@ -74,174 +76,90 @@
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
-namespace Arcane
+namespace Arcane::impl
 {
-using impl::KeyValueTextReader;
-using impl::KeyValueTextWriter;
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
-class VariableDataInfo
+VariableDataInfo::
+VariableDataInfo(const String& full_name,const ISerializedData* sdata)
+: m_full_name(full_name)
 {
- public:
-  VariableDataInfo(const String& full_name,const ISerializedData* sdata)
-  : m_full_name(full_name)
-  {
-    m_nb_dimension = sdata->nbDimension();
-    Int64ConstArrayView extents = sdata->extents();
+  m_nb_dimension = sdata->nbDimension();
+  Int64ConstArrayView extents = sdata->extents();
 
-    m_nb_base_element = sdata->nbBaseElement();
+  m_nb_base_element = sdata->nbBaseElement();
 
-    m_nb_element = sdata->nbElement();
-    m_is_multi_size = sdata->isMultiSize();
-    m_dim2_size = 0;
-    m_dim1_size = 0;
-    if (m_nb_dimension==2 && !m_is_multi_size){
-      m_dim1_size = extents[0];
-      m_dim2_size = extents[1];
-    }
-    m_dimension_array_size = extents.size();
-    m_base_data_type = sdata->baseDataType();
-    m_memory_size = sdata->memorySize();
-    m_shape = sdata->shape();
-    m_file_offset = 0;
+  m_nb_element = sdata->nbElement();
+  m_is_multi_size = sdata->isMultiSize();
+  m_dim2_size = 0;
+  m_dim1_size = 0;
+  if (m_nb_dimension==2 && !m_is_multi_size){
+    m_dim1_size = extents[0];
+    m_dim2_size = extents[1];
   }
+  m_dimension_array_size = extents.size();
+  m_base_data_type = sdata->baseDataType();
+  m_memory_size = sdata->memorySize();
+  m_shape = sdata->shape();
+  m_file_offset = 0;
+}
 
-  VariableDataInfo(const String& full_name,const XmlNode& element)
-  : m_full_name(full_name)
-  {
-    m_nb_dimension = _readInteger(element,"nb-dimension");
-    m_dim1_size = _readInt64(element,"dim1-size");
-    m_dim2_size = _readInt64(element,"dim2-size");
-    m_nb_element = _readInt64(element,"nb-element");
-    m_nb_base_element = _readInt64(element,"nb-base-element");
-    m_dimension_array_size = _readInteger(element,"dimension-array-size");
-    m_is_multi_size = _readBool(element,"is-multi-size");
-    m_base_data_type = (eDataType)_readInteger(element,"base-data-type");
-    m_memory_size = _readInt64(element,"memory-size");
-    m_file_offset = _readInt64(element,"file-offset");
-    // L'élément est nul si on repart d'une veille protection (avant Arcane 3.7)
-    XmlNode shape_attr = element.attr("shape");
-    if (!shape_attr.null()){
-      String shape_str = shape_attr.value();
-      if (!shape_str.empty()){
-        UniqueArray<Int32> values;
-        if (builtInGetValue(values,shape_str))
-          ARCANE_FATAL("Can not read values '{0}' for attribute 'shape'",shape_str);
-        m_shape.setDimensions(values);
-      }
-    }
-  }
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
 
-  const String& fullName() const { return m_full_name; }
-  Integer nbDimension() const { return m_nb_dimension; }
-  Int64 dim1Size() const { return m_dim1_size; }
-  Int64 dim2Size() const { return m_dim2_size; }
-  Int64 nbElement() const { return m_nb_element; }
-  Int64 nbBaseElement() const { return m_nb_base_element; }
-  Integer dimensionArraySize() const { return m_dimension_array_size; }
-  bool isMultiSize() const { return m_is_multi_size; }
-  eDataType baseDataType() const { return m_base_data_type; }
-  Int64 memorySize() const { return m_memory_size; }
-  const ArrayShape& shape() const { return m_shape; }
-  void setFileOffset(Int64 v){ m_file_offset = v; }
-  Int64 fileOffset() const { return m_file_offset; }
-
-  void write(XmlNode element)
-  {
-    _addAttribute(element,"nb-dimension",m_nb_dimension);
-    _addAttribute(element,"dim1-size",m_dim1_size);
-    _addAttribute(element,"dim2-size",m_dim2_size);
-    _addAttribute(element,"nb-element",m_nb_element);
-    _addAttribute(element,"nb-base-element",m_nb_base_element);
-    _addAttribute(element,"dimension-array-size",m_dimension_array_size);
-    _addAttribute(element,"is-multi-size",(m_is_multi_size) ? 1 : 0);
-    _addAttribute(element,"base-data-type",(Integer)m_base_data_type);
-    _addAttribute(element,"memory-size",m_memory_size);
-    _addAttribute(element,"file-offset",m_file_offset);
-    _addAttribute(element,"shape-size",m_shape.dimensions().size());
-    {
-      String s;
-      if (builtInPutValue(m_shape.dimensions().smallView(),s))
-        ARCANE_FATAL("Can not write '{0}'",m_shape.dimensions());
-      _addAttribute(element,"shape",s);
+VariableDataInfo::
+VariableDataInfo(const String& full_name,const XmlNode& element)
+: m_full_name(full_name)
+{
+  m_nb_dimension = _readInteger(element,"nb-dimension");
+  m_dim1_size = _readInt64(element,"dim1-size");
+  m_dim2_size = _readInt64(element,"dim2-size");
+  m_nb_element = _readInt64(element,"nb-element");
+  m_nb_base_element = _readInt64(element,"nb-base-element");
+  m_dimension_array_size = _readInteger(element,"dimension-array-size");
+  m_is_multi_size = _readBool(element,"is-multi-size");
+  m_base_data_type = (eDataType)_readInteger(element,"base-data-type");
+  m_memory_size = _readInt64(element,"memory-size");
+  m_file_offset = _readInt64(element,"file-offset");
+  // L'élément est nul si on repart d'une veille protection (avant Arcane 3.7)
+  XmlNode shape_attr = element.attr("shape");
+  if (!shape_attr.null()){
+    String shape_str = shape_attr.value();
+    if (!shape_str.empty()){
+      UniqueArray<Int32> values;
+      if (builtInGetValue(values,shape_str))
+        ARCANE_FATAL("Can not read values '{0}' for attribute 'shape'",shape_str);
+      m_shape.setDimensions(values);
     }
   }
-
- private:
-
-  void _addAttribute(XmlNode& node,const String& attr_name,Int64 value)
-  {
-    node.setAttrValue(attr_name,String::fromNumber(value));
-  }
-
-  void _addAttribute(XmlNode& node,const String& attr_name,const String& value)
-  {
-    node.setAttrValue(attr_name,value);
-  }
-
-  Integer _readInteger(const XmlNode& node,const String& attr_name)
-  {
-    return node.attr(attr_name,true).valueAsInteger(true);
-  }
-
-  Int64 _readInt64(const XmlNode& node,const String& attr_name)
-  {
-    return node.attr(attr_name,true).valueAsInt64(true);
-  }
-
-  bool _readBool(const XmlNode& node,const String& attr_name)
-  {
-    return node.attr(attr_name,true).valueAsBoolean(true);
-  }
-
-  String _readString(const XmlNode& node,const String& attr_name)
-  {
-    return node.attr(attr_name,true).value();
-  }
-
- private:
-
-  String m_full_name;
-  Integer m_nb_dimension;
-  Int64 m_dim1_size;
-  Int64 m_dim2_size;
-  Int64 m_nb_element;
-  Int64 m_nb_base_element;
-  Integer m_dimension_array_size;
-  bool m_is_multi_size;
-  eDataType m_base_data_type;
-  Int64 m_memory_size;
-  Int64 m_file_offset;
-  ArrayShape m_shape;
-};
+}
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
-class BasicVariableMetaData
+void VariableDataInfo::
+write(XmlNode element)
 {
- public:
-  BasicVariableMetaData(VariableMetaData* varmd)
+  _addAttribute(element,"nb-dimension",m_nb_dimension);
+  _addAttribute(element,"dim1-size",m_dim1_size);
+  _addAttribute(element,"dim2-size",m_dim2_size);
+  _addAttribute(element,"nb-element",m_nb_element);
+  _addAttribute(element,"nb-base-element",m_nb_base_element);
+  _addAttribute(element,"dimension-array-size",m_dimension_array_size);
+  _addAttribute(element,"is-multi-size",(m_is_multi_size) ? 1 : 0);
+  _addAttribute(element,"base-data-type",(Integer)m_base_data_type);
+  _addAttribute(element,"memory-size",m_memory_size);
+  _addAttribute(element,"file-offset",m_file_offset);
+  _addAttribute(element,"shape-size",m_shape.dimensions().size());
   {
-    m_full_name = varmd->fullName();
-    m_item_group_name = varmd->itemGroupName();
-    m_mesh_name = varmd->meshName();
-    m_item_family_name = varmd->itemFamilyName();
+    String s;
+    if (builtInPutValue(m_shape.dimensions().smallView(),s))
+      ARCANE_FATAL("Can not write '{0}'",m_shape.dimensions());
+    _addAttribute(element,"shape",s);
   }
- public:
-  const String& fullName() const { return m_full_name; }
-  const String& itemGroupName() const { return m_item_group_name; }
-  const String& meshName() const { return m_mesh_name; }
-  const String& itemFamilyName() const { return m_item_family_name; }
-  bool isItemVariable() const { return !m_item_family_name.null(); }
- private:
-  String m_full_name;
-  String m_item_group_name;
-  String m_mesh_name;
-  String m_item_family_name;
-};
+}
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
@@ -325,67 +243,34 @@ _fillUniqueIds(const ItemGroup& group,Array<Int64>& uids)
     uids.add(uid);
   }
 }
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
 }
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-/*!
- * \internal
- * \brief Lecteur générique.
- */
-class IGenericReader
+BasicGenericReader::
+BasicGenericReader(IApplication* app, Int32 version, Ref<impl::KeyValueTextReader> text_reader)
+: TraceAccessor(app->traceMng())
+, m_application(app)
+, m_text_reader(text_reader)
+, m_rank(A_NULL_RANK)
+, m_version(version)
 {
- public:
-  virtual ~IGenericReader(){}
- public:
-  virtual void initialize(const String& path,Int32 rank) =0;
-  virtual void readData(const String& var_full_name,IData* data) =0;
-  virtual void readItemGroup(const String& group_name,Int64Array& written_unique_ids,
-                             Int64Array& wanted_unique_ids) =0;
-};
+}
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
-class BasicGenericReader
-: public TraceAccessor
-, public IGenericReader
+BasicGenericReader::
+~BasicGenericReader()
 {
- public:
-  // Si 'version==-1', alors cela sera déterminé lors de
-  // l'initialisation.
-  BasicGenericReader(IApplication* app,Int32 version,Ref<KeyValueTextReader> text_reader)
-  : TraceAccessor(app->traceMng()),
-    m_application(app),
-    m_text_reader(text_reader),
-    m_rank(A_NULL_RANK),
-    m_version(version)
-  {
-  }
-  ~BasicGenericReader() override
-  {
-    for (const auto& x : m_variables_data_info )
-      delete x.second;
-  }
- public:
-  void initialize(const String& path,Int32 rank) override;
-  void readData(const String& var_full_name,IData* data) override;
-  void readItemGroup(const String& group_name,Int64Array& written_unique_ids,
-                     Int64Array& wanted_unique_ids) override;
- private:
-  IApplication* m_application;
-  Ref<KeyValueTextReader> m_text_reader;
-  String m_path;
-  Int32 m_rank;
-  Int32 m_version;
-  typedef std::map<String,VariableDataInfo*> VariableDataInfoMap;
-  VariableDataInfoMap m_variables_data_info;
- private:
-  VariableDataInfo* _getVarInfo(const String& full_name);
-};
+  for (const auto& x : m_variables_data_info)
+    delete x.second;
+}
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
@@ -582,69 +467,39 @@ readItemGroup(const String& group_full_name,Int64Array& written_unique_ids,
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-/*!
- * \internal
- * \brief Écrivain générique.
- */
-class IGenericWriter
+BasicGenericWriter::
+BasicGenericWriter(IApplication* app, Int32 version,
+                   Ref<impl::KeyValueTextWriter> text_writer)
+: TraceAccessor(app->traceMng())
+, m_application(app)
+, m_version(version)
+, m_rank(A_NULL_RANK)
+, m_text_writer(text_writer)
 {
- public:
-  virtual ~IGenericWriter(){}
- public:
-  virtual void initialize(const String& path,Int32 rank) =0;
-  virtual void writeData(const String& var_full_name,const ISerializedData* sdata) =0;
-  virtual void writeItemGroup(const String& group_full_name,
-                              SmallSpan<const Int64> written_unique_ids,
-                              SmallSpan<const Int64> wanted_unique_ids) =0;
-  virtual void endWrite() =0;
-};
+}
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
-class BasicGenericWriter
-: public TraceAccessor
-, public IGenericWriter
+BasicGenericWriter::
+~BasicGenericWriter()
 {
- public:
-  BasicGenericWriter(IApplication* app,Int32 version,
-                     Ref<KeyValueTextWriter> text_writer)
-  : TraceAccessor(app->traceMng()),
-    m_application(app),
-    m_version(version),
-    m_rank(A_NULL_RANK),
-    m_text_writer(text_writer)
-  {
+  for (const auto& x : m_variables_data_info) {
+    delete x.second;
   }
-  ~BasicGenericWriter()
-  {
-    for( const auto& x : m_variables_data_info ){
-      delete x.second;
-    }
-  }
- public:
-  void initialize(const String& path,Int32 rank) override
-  {
-    if (!m_text_writer)
-      ARCANE_FATAL("Null text writer");
-    m_path = path;
-    m_rank = rank;
-  }
-  void writeData(const String& var_full_name,const ISerializedData* sdata) override;
-  void writeItemGroup(const String& group_full_name,SmallSpan<const Int64> written_unique_ids,
-                      SmallSpan<const Int64> wanted_unique_ids) override;
-  void endWrite() override;
- private:
-  IApplication* m_application;
-  Int32 m_version;
-  String m_path;
-  Int32 m_rank;
-  Ref<KeyValueTextWriter> m_text_writer;
-  typedef std::map<String,VariableDataInfo*> VariableDataInfoMap;
-  VariableDataInfoMap m_variables_data_info;
-};
+}
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+void BasicGenericWriter::
+initialize(const String& path, Int32 rank)
+{
+  if (!m_text_writer)
+    ARCANE_FATAL("Null text writer");
+  m_path = path;
+  m_rank = rank;
+}
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
@@ -764,33 +619,6 @@ endWrite()
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
-class BasicReaderWriterCommon
-: public TraceAccessor
-{
- public:
-  enum eOpenMode
-  {
-    OpenModeRead,
-    OpenModeTruncate,
-    OpenModeAppend
-  };
- public:
-  BasicReaderWriterCommon(IApplication* app,IParallelMng* pm,
-                          const String& path,eOpenMode open_mode);
-  ~BasicReaderWriterCommon();
- protected:
-  IApplication* m_application;
-  IParallelMng* m_parallel_mng;
-  eOpenMode m_open_mode;
-  String m_path;
-  Integer m_verbose_level;
- protected:
-  String _getMetaDataFileName(Int32 rank);
-};
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
 BasicReaderWriterCommon::
 BasicReaderWriterCommon(IApplication* app,IParallelMng* pm,
                         const String& path,eOpenMode open_mode)
@@ -829,61 +657,6 @@ _getMetaDataFileName(Int32 rank)
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-/*!
- * \brief Lecture/Ecriture simple.
- */
-class BasicWriter
-: public BasicReaderWriterCommon
-, public IDataWriter
-{
- public:
-
-  BasicWriter(IApplication* app,IParallelMng* pm,const String& path,
-              eOpenMode open_mode,Integer version,bool want_parallel);
-  ~BasicWriter() override;
-
- public:
-
-  //! Positionne le service de compression. Doit être appelé avant initialize()
-  void setDataCompressor(Ref<IDataCompressor> data_compressor)
-  {
-    m_data_compressor = data_compressor;
-  }
-  void initialize();
-
-  void beginWrite(const VariableCollection& vars) override;
-  void endWrite() override;
-
-  void setMetaData(const String& meta_data) override;
-
-  void write(IVariable* v,IData* data) override;
-
- private:
-
-  bool m_want_parallel;
-  bool m_is_gather;
-  Int32 m_version;
-
-  Ref<IDataCompressor> m_data_compressor;
-  Ref<IHashAlgorithm> m_hash_algorithm;
-  Ref<KeyValueTextWriter> m_text_writer;
-
-  std::map<ItemGroup,ParallelDataWriter*> m_parallel_data_writers;
-  std::set<ItemGroup> m_written_groups;
-
-  ScopedPtrT<IGenericWriter> m_global_writer;
-
- private:
-
-  void _directWriteVal(IVariable* v,IData* data);
-  void _writeVal(impl::TextWriter* writer,VariableDataInfo* data_info,
-                 const ISerializedData* sdata);
-
-  ParallelDataWriter* _getWriter(IVariable* var);
-};
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
@@ -1117,76 +890,6 @@ endWrite()
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-/*!
- * \brief Lecteur simple.
- */
-class BasicReader
-: public BasicReaderWriterCommon
-, public IDataReader
-, public IDataReader2
-{
- public:
-  /*!
-    \brief  Interface pour retrouver le groupe associée à une variable à partir
-    de ces meta-données.
-  */
-  class IItemGroupFinder
-  {
-   public:
-    virtual ~IItemGroupFinder(){}
-    virtual ItemGroup getWantedGroup(VariableMetaData* vmd) =0;
-  };
- public:
-
-  BasicReader(IApplication* app,IParallelMng* pm,Int32 forced_rank_to_read,
-              const String& path,bool want_parallel);
-  ~BasicReader() override;
-
- public:
-
-  void beginRead(const VariableCollection& vars) override;
-  void endRead() override {}
-  String metaData() override;
-  void read(IVariable* v,IData* data) override;
-
-  void fillMetaData(ByteArray& bytes) override;
-  void beginRead(const DataReaderInfo& infos) override;
-  void read(const VariableDataReadInfo& infos) override;
-
- public:
-
-  void initialize();
-  void setItemGroupFinder(IItemGroupFinder* group_finder)
-  {
-    m_item_group_finder = group_finder;
-  }
- private:
-
-  bool m_want_parallel;
-  Integer m_nb_written_part;
-  Int32 m_version;
-
-  Int32 m_first_rank_to_read;
-  Int32 m_nb_rank_to_read;
-  Int32 m_forced_rank_to_read;
-
-  std::map<String,ParallelDataReader*> m_parallel_data_readers;
-  UniqueArray<IGenericReader*> m_global_readers;
-  IItemGroupFinder* m_item_group_finder;
-  Ref<KeyValueTextReader> m_forced_rank_to_read_text_reader; //!< Lecteur pour le premier rang à lire.
-  Ref<IDataCompressor> m_data_compressor;
-
- private:
-
-  void _directReadVal(VariableMetaData* varmd,IData* data);
-
-  ParallelDataReader* _getReader(VariableMetaData* varmd);
-  void _setRanksToRead();
-  IGenericReader* _readOwnMetaDataAndCreateReader(Int32 rank);
-};
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
@@ -1592,8 +1295,14 @@ beginRead(const DataReaderInfo& infos)
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
+}
+
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
+
+namespace Arcane
+{
+using namespace Arcane::impl;
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
